// src/shared/api/models/getApiInsightsStockReports200.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiInsightsStockReports200Data } from './getApiInsightsStockReports200Data';

export type GetApiInsightsStockReports200 = {
    data: GetApiInsightsStockReports200Data;
    message: string;
};