// src/shared/api/models/getApiInsightsStockReportsSortBy.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiInsightsStockReportsSortBy =
    (typeof GetApiInsightsStockReportsSortBy)[keyof typeof GetApiInsightsStockReportsSortBy];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiInsightsStockReportsSortBy = {
    name: 'name',
    brand: 'brand',
    salePrice: 'salePrice',
    purchasePrice: 'purchasePrice',
    stockStatus: 'stockStatus',
    inventoryValue: 'inventoryValue',
} as const;